Rename dcp_check_report module
